pyproject.toml
[tool.poetry]
name = "twiiter-etl"
version = "0.1.1"
description = "Twiiter ETL Pipeline"
authors = ["Hamza Ali"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
tweepy = "^4.14.0"
s3fs = "^2024.6.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
